requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.30.0

django>2.0,<2.1.0

# Configuration
django-environ==0.4.4

whitenoise==3.3.1


# Forms
django-braces>=1.13.0
django-crispy-forms>=1.7.0
django-floppyforms==1.7.0



# Models
django-model-utils==3.1.1

# Images
Pillow==5.0.0

# Password storage
argon2-cffi==18.1.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.38.0


# Python-PostgreSQL Database Adapter
psycopg2==2.7.3.2

# Unicode slugification
awesome-slugify==1.6.5

# Time zones support
pytz==2017.3

# Redis support
django-redis==4.8.0
redis>=2.10.5


# Scheduled
celery==4.0.2
django-celery-results>=1.0.1
django-celery-beat>=1.0.1



rcssmin==1.0.6 
django-compressor==2.2


### PURE CREATIVE ADDITIONS



django-extensions==1.9.9

# Django Common
-e git+ssh://git@github.com/sam-mi/django-common.git@master#egg=django-common
# django-typed-models==0.8.0

# Django Closure Tree does note yet support django 2, there is a pull request for it:
# gitversion==2.1.3
# -e git+https://github.com/jsmitka/django-closuretree.git@django-2.0#egg=django-closuretree
# django-closuretree==1.2.0

# CMS



Markdown





#Admin
django-admin-tools==0.8.1
# select2 and raw_ids
-e git+https://github.com/arthanson/django-genericadmin.git@master#egg=django-genericadmin

django-guardian==1.4.9




django-jsonfield-compat==0.4.4

django-markupfield==1.4.2




# FILTERS
django-mathfilters


django-ipware==1.1.
django-flat-theme==1.1.4

lxml>=3.5.0


